+++ lib/solargraph/source_map/mapper.rb
@@ -63,7 +63,7 @@ module Solargraph
       end
 
       def process_comment source_position, comment_position, comment
-        return unless comment =~ MACRO_REGEXP
+        return unless comment.encode('UTF-8', invalid: :replace, replace: '?') =~ MACRO_REGEXP
         cmnt = remove_inline_comment_hashes(comment)
         parse = Solargraph::Source.parse_docstring(cmnt)
         last_line = 0
@@ -171,7 +171,7 @@ module Solargraph
             # @todo Handle parser errors in !parse directives
           end
         when 'domain'
-          namespace = closure_at(source_position)
+          namespace = closure_at(source_position) || Pin::ROOT_PIN
           namespace.domains.concat directive.tag.types unless directive.tag.types.nil?
         when 'override'
           pins.push Pin::Reference::Override.new(location, directive.tag.name, docstring.tags)
@@ -184,7 +184,7 @@ module Solargraph
         started = false
         comment.lines.each { |l|
           # Trim the comment and minimum leading whitespace
-          p = l.gsub(/^#/, '')
+          p = l.encode('UTF-8', invalid: :replace, replace: '?').gsub(/^#/, '')
           if num.nil? && !p.strip.empty?
             num = p.index(/[^ ]/)
             started = true
@@ -199,7 +199,7 @@ module Solargraph
 
       # @return [void]
       def process_comment_directives
-        return unless @code =~ MACRO_REGEXP
+        return unless @code.encode('UTF-8', invalid: :replace, replace: '?') =~ MACRO_REGEXP
         code_lines = @code.lines
         @source.associated_comments.each do |line, comments|
           src_pos = line ? Position.new(line, code_lines[line].to_s.chomp.index(/[^\s]/) || 0) : Position.new(code_lines.length, 0)