+++ spec/source_map/mapper_spec.rb
@@ -1508,4 +1508,16 @@ describe Solargraph::SourceMap::Mapper do
     ).map(&:name)
     expect(locals).to eq(['lvar'])
   end
+
+  it 'handles invalid byte sequences' do
+    expect {
+      Solargraph::SourceMap.load('spec/fixtures/invalid_byte.rb')
+    }.not_to raise_error
+  end
+
+  it 'handles invalid byte sequences in stringified node comments' do
+    expect {
+      Solargraph::SourceMap.load('spec/fixtures/invalid_node_comment.rb')
+    }.not_to raise_error
+  end
 end